LayoutTests/http/tests/media/media-stream/enumerate-devices-source-id-expected.txt

 
 

Tests that mediaDevices.enumerateDevices returns a unique ID for deviceId in each domain.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS : device id 1 is unique
PASS : device id 2 is unique
PASS : device id 3 is unique
PASS : device id 4 is unique
PASS : device id 5 is unique
PASS : device id 6 is unique
PASS : device id 7 is unique
PASS : device id 8 is unique
PASS : device id 9 is unique
PASS : device id 10 is unique
PASS : device id 11 is unique
PASS : device id 12 is unique

PASS : all device IDs are unique

PASS successfullyParsed is true

TEST COMPLETE
